Revert "code exchange - /templates revamp (#2358)"
This reverts commit 58f127ed4336651ae46f9573d6f237acbda5c439.
